requirements.txt
dj-database-url==0.4.2
Django==1.11.2
djangorestframework==3.6.3
gunicorn==19.7.1
pytz==2017.2
